requirements.txt
Cython
datasets
timm
wandb
tensorboard
lightning
schedulefree
einops
numpy
matplotlib
accelerate
diffusers
torchmetrics
hf_transfer
Cython
mosaicml-streaming